cursor_pdf_import_kit/requirements.txt
pdfminer.six==20231228
pypdf==4.3.1
